+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/ags/BagsStatusListener.java
@@ -1,0 +1,9 @@
+package jsky.app.ot.ags;
+
+import edu.gemini.pot.sp.ISPObservation;
+import edu.gemini.shared.util.immutable.Option;
+
+@FunctionalInterface
+public interface BagsStatusListener {
+    void bagsStatusChanged(final ISPObservation obs, final Option<BagsStatus> oldStatus, final Option<BagsStatus> newStatus);
+}
