+++ bundle/jsky.app.ot/src/main/scala/jsky/app/ot/ags/BagsStatus.scala
@@ -6,5 +6,4 @@ object BagsStatus {
   case object Pending extends BagsStatus("BAGS lookup pending.")
   case object Running extends BagsStatus("BAGS running...")
   case class Failed(reason: String)  extends BagsStatus(s"BAGS failed to execute: $reason")
-  case object NoTarget extends BagsStatus("BAGS could not find a target.")
 }
